Feat: Tag Ranking API 구현

+++ back-end/backend/src/main/java/com/groot/backend/entity/TagCountEntity.java
@@ -21,5 +21,5 @@ public class TagCountEntity {
     @Column(unique = true)
     private String tag;
     @Column
-    private Integer count;
+    private Double count;
 }
